refactor(clipool): extract worker pure helpers to _worker_helpers — file_length gate (#1009)
